Removed fail2ban support on Mac OS
Fail2ban package in Home-brew appears broken. At the very least I couldn’t get it to work. Ultimately could not get IP’s banned through PF to be blocked. The nature of Mac OS means this is less important as they’re usually on a private network. Learning PF and fixing homebrew package is beyond the scope of this project.